Set ITSAppUsesNonExemptEncryption to false
This commit updates the Info.plist to include the `ITSAppUsesNonExemptEncryption` key and sets its value to false.

This declaration is required for App Store submission and simplifies the process by indicating that the app does not use non-exempt encryption, which helps to avoid extra steps related to export compliance documentation.